Shift the focus of this package a little.
We now re-use the UID catalog metadata/column that is already used for Archetypes content. We still register an indexer for this, which will work with any IUUIDAware content, but we don't need to install a new catalog index anymore.

The rationale behind this is that Archetypes 1.7 (or at least the plip10778-plone.uuid branch, which will hopefully be part of 1.7) now has its own plone.uuid integration, using it to generate new UUIDs and return UUIDs from its UID() function (which basically just becomes a BBB shim around IUUID(object)). By using the existing catalog setup, we can avoid Archetypes migration and make compatibility with old code feasible.

Refs #10778.

svn path=/plone.app.uuid/trunk/; revision=39393