Speed up uuidToPhysicalPath and uuidToObject.
Do this by using an IndexQuery to only query the UID index.
This means we do not need to get a full brain, which takes long.

Before, in a `test_speed` run with 400 items:

```
Time taken for uuidToPhysicalPath: 0.21054577827453613
Time taken for uuidToObject: 0.4477109909057617
```

Afterwards:

```
Time taken for uuidToPhysicalPath: 0.030041933059692383
Time taken for uuidToObject: 0.09154772758483887
```

So this is 5 to 7 times as fast.

I tried changing uuidToURL along the same lines, but in both cases this was about the same, 0.15 seconds.